moved appl logic to sub component
